refactor: 볼륨이 큰 when, where 폴더 분리, 동작하지 않는 svgr 속성 제거 로직 제거
