+++ svgr.d.ts
@@ -1,7 +1,13 @@
 // svgr 폴더는 React 컴포넌트 취급
 declare module '@/assets/svgs/**/*.svg' {
   import * as React from 'react';
-  const SVGComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+  const SVGComponent: React.FunctionComponent<
+    React.SVGProps<SVGSVGElement> & {
+      className?: string,
+      width?: number | string,
+      height?: number | string,
+    }
+  >;
   export default SVGComponent;
 }
 // declare module '*/assets/svgs/**/*.svg' {
@@ -16,7 +22,7 @@ declare module '@/assets/svgs/**/*.svg' {
 
 declare module '*.svg' {
   import * as React from 'react';
-  const SVGComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+  const SVGComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement> & {className?: string}>;
   export default SVGComponent;
   export const ReactComponent: typeof SVGComponent;
 }
